third_party/blink/web_tests/fragmentation/repeating-thead-exceeds-page-size.html

<!DOCTYPE html>
<style>
table {
    border-collapse: collapse;
}
td, th {
}
thead, tr {
    break-inside: avoid;
}
</style>
<p>crbug.com/621258: Drop the repeated header if there isn't room for at least one content row on the page.</p>
<div style="-webkit-columns:3; line-height: 20px; column-fill: auto; height:190px;">
    <table id="top">
        <thead id="top-thead">
            <tr>
                <th><div style="height: 195px;">Col 1</div></th>
            </tr>
        </thead>
        <tr><td>Te</td><td>xt</td></tr>
        <tr><td>Te</td><td>xt</td></tr>
        <tr><td>Te</td><td>xt</td></tr>
        <tr><td>Te</td><td>xt</td></tr>
        <tr><td>Te</td><td>xt</td></tr>
        <tr><td>Te</td><td>xt</td></tr>
        <tr><td>Te</td><td>xt</td></tr>
        <tr><td>Te</td><td>xt</td></tr>
        <tr><td class="cell" data-total-y=373>Te</td><td>xt</td></tr>
        <tr><td>Te</td><td>xt</td></tr>
        <tr><td>Te</td><td>xt</td></tr>
        <tr><td>Te</td><td>xt</td></tr>
        <tr><td>Te</td><td>xt</td></tr>
        <tr><td>Te</td><td>xt</td></tr>
        <tr><td>Te</td><td>xt</td></tr>
        <tr><td>Te</td><td>xt</td></tr>
        <tr><td class="cell" data-total-y=549>Te</td><td>xt</td></tr>
        <tr><td>Te</td><td>xt</td></tr>
    </table>
</div>
<div id="console"></div>
<script src="../resources/check-layout.js"></script>
<script>
checkLayout(".cell", document.getElementById("console"));
</script>
