third_party/blink/web_tests/platform/generic/fragmentation/repeating-thead-exceeds-page-size-expected.txt

crbug.com/621258: Drop the repeated header if there isn't room for at least one content row on the page.

Col 1

Te	xt
Te	xt
Te	xt
Te	xt
Te	xt
Te	xt
Te	xt
Te	xt
Te	xt
Te	xt
Te	xt
Te	xt
Te	xt
Te	xt
Te	xt
Te	xt
Te	xt
Te	xt
PASS
PASS
